# Fixes for PHP 8.2 crashes