--- a/ext/include/arithmetic.c
+++ b/ext/include/arithmetic.c
@@ -14,15 +14,12 @@ void tensor_multiply(zval * return_value, zval * a, zval * b)
     zend_array * aa = Z_ARR_P(a);
     zend_array * ab = Z_ARR_P(b);
 
-    Bucket * ba = aa->arData;
-    Bucket * bb = ab->arData;
-
     unsigned int n = zend_array_count(aa);
 
 	array_init_size(&c, n);
 
     for (i = 0; i < n; ++i) {
-        product = zephir_get_doubleval(&ba[i].val) * zephir_get_doubleval(&bb[i].val);
+        product = zephir_get_doubleval(zend_hash_index_find(aa, i)) * zephir_get_doubleval(zend_hash_index_find(ab, i));
 
 	    add_next_index_double(&c, product);
     }
@@ -39,15 +36,12 @@ void tensor_divide(zval * return_value, zval * a, zval * b)
     zend_array * aa = Z_ARR_P(a);
     zend_array * ab = Z_ARR_P(b);
 
-    Bucket * ba = aa->arData;
-    Bucket * bb = ab->arData;
-
     unsigned int n = zend_array_count(aa);
 
 	array_init_size(&c, n);
 
     for (i = 0; i < n; ++i) {
-        quotient = zephir_get_doubleval(&ba[i].val) / zephir_get_doubleval(&bb[i].val);
+        quotient = zephir_get_doubleval(zend_hash_index_find(aa, i)) / zephir_get_doubleval(zend_hash_index_find(ab, i));
 
 	    add_next_index_double(&c, quotient);
     }
@@ -64,15 +58,12 @@ void tensor_add(zval * return_value, zval * a, zval * b)
     zend_array * aa = Z_ARR_P(a);
     zend_array * ab = Z_ARR_P(b);
 
-    Bucket * ba = aa->arData;
-    Bucket * bb = ab->arData;
-
     unsigned int n = zend_array_count(aa);
 
 	array_init_size(&c, n);
 
     for (i = 0; i < n; ++i) {
-        sum = zephir_get_doubleval(&ba[i].val) + zephir_get_doubleval(&bb[i].val);
+        sum = zephir_get_doubleval(zend_hash_index_find(aa, i)) + zephir_get_doubleval(zend_hash_index_find(ab, i));
 
 	    add_next_index_double(&c, sum);
     }
@@ -89,15 +80,12 @@ void tensor_subtract(zval * return_value, zval * a, zval * b)
     zend_array * aa = Z_ARR_P(a);
     zend_array * ab = Z_ARR_P(b);
 
-    Bucket * ba = aa->arData;
-    Bucket * bb = ab->arData;
-
     unsigned int n = zend_array_count(aa);
 
 	array_init_size(&c, n);
 
     for (i = 0; i < n; ++i) {
-        difference = zephir_get_doubleval(&ba[i].val) - zephir_get_doubleval(&bb[i].val);
+        difference = zephir_get_doubleval(zend_hash_index_find(aa, i)) - zephir_get_doubleval(zend_hash_index_find(ab, i));
 
 	    add_next_index_double(&c, difference);
     }
@@ -114,15 +102,12 @@ void tensor_pow(zval * return_value, zval * a, zval * b)
     zend_array * aa = Z_ARR_P(a);
     zend_array * ab = Z_ARR_P(b);
 
-    Bucket * ba = aa->arData;
-    Bucket * bb = ab->arData;
-
     unsigned int n = zend_array_count(aa);
 
 	array_init_size(&c, n);
 
     for (i = 0; i < n; ++i) {
-        power = pow(zephir_get_doubleval(&ba[i].val), zephir_get_doubleval(&bb[i].val));
+        power = pow(zephir_get_doubleval(zend_hash_index_find(aa, i)), zephir_get_doubleval(zend_hash_index_find(ab, i)));
 
 	    add_next_index_double(&c, power);
     }
@@ -139,15 +124,12 @@ void tensor_mod(zval * return_value, zval * a, zval * b)
     zend_array * aa = Z_ARR_P(a);
     zend_array * ab = Z_ARR_P(b);
 
-    Bucket * ba = aa->arData;
-    Bucket * bb = ab->arData;
-
     unsigned int n = zend_array_count(aa);
 
 	array_init_size(&c, n);
 
     for (i = 0; i < n; ++i) {
-        mod_function(&modulus, &ba[i].val, &bb[i].val);
+        mod_function(&modulus, zend_hash_index_find(aa, i), zend_hash_index_find(ab, i));
 
 	    add_next_index_zval(&c, &modulus);
     }
@@ -163,8 +145,6 @@ void tensor_multiply_scalar(zval * return_value, zval * a, zval * b)
 
     zend_array * aa = Z_ARR_P(a);
 
-    Bucket * ba = aa->arData;
-
     double ab = zephir_get_doubleval(b);
 
     unsigned int n = zend_array_count(aa);
@@ -172,7 +152,7 @@ void tensor_multiply_scalar(zval * return_value, zval * a, zval * b)
 	array_init_size(&c, n);
 
     for (i = 0; i < n; ++i) {
-        product = zephir_get_doubleval(&ba[i].val) * ab;
+        product = zephir_get_doubleval(zend_hash_index_find(aa, i)) * ab;
 
 	    add_next_index_double(&c, product);
     }
@@ -188,8 +168,6 @@ void tensor_divide_scalar(zval * return_value, zval * a, zval * b)
 
     zend_array * aa = Z_ARR_P(a);
 
-    Bucket * ba = aa->arData;
-
     double ab = zephir_get_doubleval(b);
 
     unsigned int n = zend_array_count(aa);
@@ -197,7 +175,7 @@ void tensor_divide_scalar(zval * return_value, zval * a, zval * b)
 	array_init_size(&c, n);
 
     for (i = 0; i < n; ++i) {
-        quotient = zephir_get_doubleval(&ba[i].val) / ab;
+        quotient = zephir_get_doubleval(zend_hash_index_find(aa, i)) / ab;
 
 	    add_next_index_double(&c, quotient);
     }
@@ -213,8 +191,6 @@ void tensor_add_scalar(zval * return_value, zval * a, zval * b)
 
     zend_array * aa = Z_ARR_P(a);
 
-    Bucket * ba = aa->arData;
-
     double ab = zephir_get_doubleval(b);
 
     unsigned int n = zend_array_count(aa);
@@ -222,7 +198,7 @@ void tensor_add_scalar(zval * return_value, zval * a, zval * b)
 	array_init_size(&c, n);
 
     for (i = 0; i < n; ++i) {
-        sum = zephir_get_doubleval(&ba[i].val) + ab;
+        sum = zephir_get_doubleval(zend_hash_index_find(aa, i)) + ab;
 
 	    add_next_index_double(&c, sum);
     }
@@ -238,8 +214,6 @@ void tensor_subtract_scalar(zval * return_value, zval * a, zval * b)
 
     zend_array * aa = Z_ARR_P(a);
 
-    Bucket * ba = aa->arData;
-
     double ab = zephir_get_doubleval(b);
 
     unsigned int n = zend_array_count(aa);
@@ -247,7 +221,7 @@ void tensor_subtract_scalar(zval * return_value, zval * a, zval * b)
 	array_init_size(&c, n);
 
     for (i = 0; i < n; ++i) {
-        difference = zephir_get_doubleval(&ba[i].val) - ab;
+        difference = zephir_get_doubleval(zend_hash_index_find(aa, i)) - ab;
 
 	    add_next_index_double(&c, difference);
     }
@@ -263,8 +237,6 @@ void tensor_pow_scalar(zval * return_value, zval * a, zval * b)
 
     zend_array * aa = Z_ARR_P(a);
 
-    Bucket * ba = aa->arData;
-
     double ab = zephir_get_doubleval(b);
 
     unsigned int n = zend_array_count(aa);
@@ -272,7 +244,7 @@ void tensor_pow_scalar(zval * return_value, zval * a, zval * b)
 	array_init_size(&c, n);
 
     for (i = 0; i < n; ++i) {
-        power = pow(zephir_get_doubleval(&ba[i].val), ab);
+        power = pow(zephir_get_doubleval(zend_hash_index_find(aa, i)), ab);
 
 	    add_next_index_double(&c, power);
     }
@@ -288,14 +260,12 @@ void tensor_mod_scalar(zval * return_value, zval * a, zval * b)
 
     zend_array * aa = Z_ARR_P(a);
 
-    Bucket * ba = aa->arData;
-
     unsigned int n = zend_array_count(aa);
 
 	array_init_size(&c, n);
 
     for (i = 0; i < n; ++i) {
-        mod_function(&modulus, &ba[i].val, b);
+        mod_function(&modulus, zend_hash_index_find(aa, i), b);
 
 	    add_next_index_zval(&c, &modulus);
     }