--- a/ext/include/comparison.c
+++ b/ext/include/comparison.c
@@ -13,15 +13,12 @@ void tensor_equal(zval * return_value, zval * a, zval * b)
     zend_array * aa = Z_ARR_P(a);
     zend_array * ab = Z_ARR_P(b);
 
-    Bucket * ba = aa->arData;
-    Bucket * bb = ab->arData;
-
     unsigned int n = zend_array_count(aa);
 
 	array_init_size(&c, n);
 
     for (i = 0; i < n; ++i) {
-        if (zephir_get_doubleval(&ba[i].val) == zephir_get_doubleval(&bb[i].val)) {
+        if (zephir_get_doubleval(zend_hash_index_find(aa, i)) == zephir_get_doubleval(zend_hash_index_find(ab, i))) {
             add_next_index_long(&c, 1);
         } else {
             add_next_index_long(&c, 0);
@@ -39,15 +36,12 @@ void tensor_not_equal(zval * return_value, zval * a, zval * b)
     zend_array * aa = Z_ARR_P(a);
     zend_array * ab = Z_ARR_P(b);
 
-    Bucket * ba = aa->arData;
-    Bucket * bb = ab->arData;
-
     unsigned int n = zend_array_count(aa);
 
 	array_init_size(&c, n);
 
     for (i = 0; i < n; ++i) {
-        if (zephir_get_doubleval(&ba[i].val) != zephir_get_doubleval(&bb[i].val)) {
+        if (zephir_get_doubleval(zend_hash_index_find(aa, i)) != zephir_get_doubleval(zend_hash_index_find(ab, i))) {
             add_next_index_long(&c, 1);
         } else {
             add_next_index_long(&c, 0);
@@ -65,15 +59,12 @@ void tensor_greater(zval * return_value, zval * a, zval * b)
     zend_array * aa = Z_ARR_P(a);
     zend_array * ab = Z_ARR_P(b);
 
-    Bucket * ba = aa->arData;
-    Bucket * bb = ab->arData;
-
     unsigned int n = zend_array_count(aa);
 
 	array_init_size(&c, n);
 
     for (i = 0; i < n; ++i) {
-        if (zephir_get_doubleval(&ba[i].val) > zephir_get_doubleval(&bb[i].val)) {
+        if (zephir_get_doubleval(zend_hash_index_find(aa, i)) > zephir_get_doubleval(zend_hash_index_find(ab, i))) {
             add_next_index_long(&c, 1);
         } else {
             add_next_index_long(&c, 0);
@@ -91,15 +82,12 @@ void tensor_greater_equal(zval * return_value, zval * a, zval * b)
     zend_array * aa = Z_ARR_P(a);
     zend_array * ab = Z_ARR_P(b);
 
-    Bucket * ba = aa->arData;
-    Bucket * bb = ab->arData;
-
     unsigned int n = zend_array_count(aa);
 
 	array_init_size(&c, n);
 
     for (i = 0; i < n; ++i) {
-        if (zephir_get_doubleval(&ba[i].val) >= zephir_get_doubleval(&bb[i].val)) {
+        if (zephir_get_doubleval(zend_hash_index_find(aa, i)) >= zephir_get_doubleval(zend_hash_index_find(ab, i))) {
             add_next_index_long(&c, 1);
         } else {
             add_next_index_long(&c, 0);
@@ -117,15 +105,12 @@ void tensor_less(zval * return_value, zval * a, zval * b)
     zend_array * aa = Z_ARR_P(a);
     zend_array * ab = Z_ARR_P(b);
 
-    Bucket * ba = aa->arData;
-    Bucket * bb = ab->arData;
-
     unsigned int n = zend_array_count(aa);
 
 	array_init_size(&c, n);
 
     for (i = 0; i < n; ++i) {
-        if (zephir_get_doubleval(&ba[i].val) < zephir_get_doubleval(&bb[i].val)) {
+        if (zephir_get_doubleval(zend_hash_index_find(aa, i)) < zephir_get_doubleval(zend_hash_index_find(ab, i))) {
             add_next_index_long(&c, 1);
         } else {
             add_next_index_long(&c, 0);
@@ -143,15 +128,12 @@ void tensor_less_equal(zval * return_value, zval * a, zval * b)
     zend_array * aa = Z_ARR_P(a);
     zend_array * ab = Z_ARR_P(b);
 
-    Bucket * ba = aa->arData;
-    Bucket * bb = ab->arData;
-
     unsigned int n = zend_array_count(aa);
 
 	array_init_size(&c, n);
 
     for (i = 0; i < n; ++i) {
-        if (zephir_get_doubleval(&ba[i].val) <= zephir_get_doubleval(&bb[i].val)) {
+        if (zephir_get_doubleval(zend_hash_index_find(aa, i)) <= zephir_get_doubleval(zend_hash_index_find(ab, i))) {
             add_next_index_long(&c, 1);
         } else {
             add_next_index_long(&c, 0);
@@ -168,8 +150,6 @@ void tensor_equal_scalar(zval * return_value, zval * a, zval * b)
 
     zend_array * aa = Z_ARR_P(a);
 
-    Bucket * ba = aa->arData;
-
     double ab = zephir_get_doubleval(b);
 
     unsigned int n = zend_array_count(aa);
@@ -177,7 +157,7 @@ void tensor_equal_scalar(zval * return_value, zval * a, zval * b)
 	array_init_size(&c, n);
 
     for (i = 0; i < n; ++i) {
-        if (zephir_get_doubleval(&ba[i].val) == ab) {
+        if (zephir_get_doubleval(zend_hash_index_find(aa, i)) == ab) {
             add_next_index_long(&c, 1);
         } else {
             add_next_index_long(&c, 0);
@@ -194,8 +174,6 @@ void tensor_not_equal_scalar(zval * return_value, zval * a, zval * b)
 
     zend_array * aa = Z_ARR_P(a);
 
-    Bucket * ba = aa->arData;
-
     double ab = zephir_get_doubleval(b);
 
     unsigned int n = zend_array_count(aa);
@@ -203,7 +181,7 @@ void tensor_not_equal_scalar(zval * return_value, zval * a, zval * b)
 	array_init_size(&c, n);
 
     for (i = 0; i < n; ++i) {
-        if (zephir_get_doubleval(&ba[i].val) != ab) {
+        if (zephir_get_doubleval(zend_hash_index_find(aa, i)) != ab) {
             add_next_index_long(&c, 1);
         } else {
             add_next_index_long(&c, 0);
@@ -220,8 +198,6 @@ void tensor_greater_scalar(zval * return_value, zval * a, zval * b)
 
     zend_array * aa = Z_ARR_P(a);
 
-    Bucket * ba = aa->arData;
-
     double ab = zephir_get_doubleval(b);
 
     unsigned int n = zend_array_count(aa);
@@ -229,7 +205,7 @@ void tensor_greater_scalar(zval * return_value, zval * a, zval * b)
 	array_init_size(&c, n);
 
     for (i = 0; i < n; ++i) {
-        if (zephir_get_doubleval(&ba[i].val) > ab) {
+        if (zephir_get_doubleval(zend_hash_index_find(aa, i)) > ab) {
             add_next_index_long(&c, 1);
         } else {
             add_next_index_long(&c, 0);
@@ -246,8 +222,6 @@ void tensor_greater_equal_scalar(zval * return_value, zval * a, zval * b)
 
     zend_array * aa = Z_ARR_P(a);
 
-    Bucket * ba = aa->arData;
-
     double ab = zephir_get_doubleval(b);
 
     unsigned int n = zend_array_count(aa);
@@ -255,7 +229,7 @@ void tensor_greater_equal_scalar(zval * return_value, zval * a, zval * b)
 	array_init_size(&c, n);
 
     for (i = 0; i < n; ++i) {
-        if (zephir_get_doubleval(&ba[i].val) >= ab) {
+        if (zephir_get_doubleval(zend_hash_index_find(aa, i)) >= ab) {
             add_next_index_long(&c, 1);
         } else {
             add_next_index_long(&c, 0);
@@ -272,8 +246,6 @@ void tensor_less_scalar(zval * return_value, zval * a, zval * b)
 
     zend_array * aa = Z_ARR_P(a);
 
-    Bucket * ba = aa->arData;
-
     double ab = zephir_get_doubleval(b);
 
     unsigned int n = zend_array_count(aa);
@@ -281,7 +253,7 @@ void tensor_less_scalar(zval * return_value, zval * a, zval * b)
 	array_init_size(&c, n);
 
     for (i = 0; i < n; ++i) {
-        if (zephir_get_doubleval(&ba[i].val) < ab) {
+        if (zephir_get_doubleval(zend_hash_index_find(aa, i)) < ab) {
             add_next_index_long(&c, 1);
         } else {
             add_next_index_long(&c, 0);
@@ -307,7 +279,7 @@ void tensor_less_equal_scalar(zval * return_value, zval * a, zval * b)
 	array_init_size(&c, n);
 
     for (i = 0; i < n; ++i) {
-        if (zephir_get_doubleval(&ba[i].val) <= ab) {
+        if (zephir_get_doubleval(zend_hash_index_find(aa, i)) <= ab) {
             add_next_index_long(&c, 1);
         } else {
             add_next_index_long(&c, 0);