--- a/ext/include/linear_algebra.c
+++ b/ext/include/linear_algebra.c
@@ -17,36 +17,33 @@
 void tensor_matmul(zval * return_value, zval * a, zval * b)
 {
     unsigned int i, j;
-    Bucket * row;
+    zval * row;
     zval rowC, c;
 
     zend_array * aa = Z_ARR_P(a);
     zend_array * ab = Z_ARR_P(b);
 
-    Bucket * ba = aa->arData;
-    Bucket * bb = ab->arData;
-
     unsigned int m = zend_array_count(aa);
     unsigned int p = zend_array_count(ab);
-    unsigned int n = zend_array_count(Z_ARR(bb[0].val));
+    unsigned int n = zend_array_count(Z_ARR_P(zend_hash_index_find(ab, 0)));
 
     double * va = emalloc(m * p * sizeof(double));
     double * vb = emalloc(n * p * sizeof(double));
     double * vc = emalloc(m * n * sizeof(double));
 
     for (i = 0; i < m; ++i) {
-        row = Z_ARR(ba[i].val)->arData;
+        row = zend_hash_index_find(aa, i);
 
         for (j = 0; j < p; ++j) {
-            va[i * p + j] = zephir_get_doubleval(&row[j].val);
+            va[i * p + j] = zephir_get_doubleval(zend_hash_index_find(Z_ARR_P(row), j));
         }
     }
 
     for (i = 0; i < p; ++i) {
-        row = Z_ARR(bb[i].val)->arData;
+        row = zend_hash_index_find(ab, i);
 
         for (j = 0; j < n; ++j) {
-            vb[i * n + j] = zephir_get_doubleval(&row[j].val);
+            vb[i * n + j] = zephir_get_doubleval(zend_hash_index_find(Z_ARR_P(row), j));
         }
     }
 
@@ -85,15 +82,12 @@ void tensor_dot(zval * return_value, zval * a, zval * b)
     zend_array * aa = Z_ARR_P(a);
     zend_array * ab = Z_ARR_P(b);
 
-    Bucket * ba = aa->arData;
-    Bucket * bb = ab->arData;
-
     unsigned int n = zend_array_count(aa);
 
     double sigma = 0.0;
 
     for (i = 0; i < n; ++i) {
-        sigma += zephir_get_doubleval(&ba[i].val) * zephir_get_doubleval(&bb[i].val);
+        sigma += zephir_get_doubleval(zend_hash_index_find(aa, i)) * zephir_get_doubleval(zend_hash_index_find(ab, i));
     }
 
     RETVAL_DOUBLE(sigma);
@@ -101,19 +95,17 @@ void tensor_dot(zval * return_value, zval * a, zval * b)
 
 /**
  * Return the multiplicative inverse of a square matrix A.
- * 
+ *
  * @param return_value
  * @param a
  */
 void tensor_inverse(zval * return_value, zval * a)
 {
     unsigned int i, j;
-    Bucket * row;
+    zval * row;
     zval rowB, b;
 
     zend_array * aa = Z_ARR_P(a);
-
-    Bucket * ba = aa->arData;
 
     unsigned int n = zend_array_count(aa);
 
@@ -121,13 +113,13 @@ void tensor_inverse(zval * return_value, zval * a)
     int * pivots = emalloc(n * sizeof(int));
 
     for (i = 0; i < n; ++i) {
-        row = Z_ARR(ba[i].val)->arData;
+        row = zend_hash_index_find(aa, i);
 
         for (j = 0; j < n; ++j) {
-            va[i * n + j] = zephir_get_doubleval(&row[j].val);
+            va[i * n + j] = zephir_get_doubleval(zend_hash_index_find(Z_ARR_P(row), j));
         }
     }
-
+    
     lapack_int status;
 
     status = LAPACKE_dgetrf(LAPACK_ROW_MAJOR, n, n, va, n, pivots);
@@ -169,15 +161,13 @@ void tensor_inverse(zval * return_value, zval * a)
 void tensor_pseudoinverse(zval * return_value, zval * a)
 {
     unsigned int i, j;
-    Bucket * row;
+    zval * row;
     zval b, rowB;
 
     zend_array * aa = Z_ARR_P(a);
 
-    Bucket * ba = aa->arData;
-
     unsigned int m = zend_array_count(aa);
-    unsigned int n = zend_array_count(Z_ARR(ba[0].val));
+    unsigned int n = zend_array_count(Z_ARR_P(zend_hash_index_find(aa, 0)));
     unsigned int k = MIN(m, n);
 
     double * va = emalloc(m * n * sizeof(double));
@@ -187,10 +177,10 @@ void tensor_pseudoinverse(zval * return_value, zval * a)
     double * vb = emalloc(n * m * sizeof(double));
 
     for (i = 0; i < m; ++i) {
-        row = Z_ARR(ba[i].val)->arData;
+        row = zend_hash_index_find(aa, i);
 
         for (j = 0; j < n; ++j) {
-            va[i * n + j] = zephir_get_doubleval(&row[j].val);
+            va[i * n + j] = zephir_get_doubleval(zend_hash_index_find(Z_ARR_P(row), j));
         }
     }
 
@@ -236,25 +226,23 @@ void tensor_pseudoinverse(zval * return_value, zval * a)
 void tensor_ref(zval * return_value, zval * a)
 {
     unsigned int i, j;
-    Bucket * row;
+    zval * row;
     zval rowB, b;
     zval tuple;
 
     zend_array * aa = Z_ARR_P(a);
 
-    Bucket * ba = aa->arData;
-
     unsigned int m = zend_array_count(aa);
-    unsigned int n = zend_array_count(Z_ARR(ba[0].val));
+    unsigned int n = zend_array_count(Z_ARR_P(zend_hash_index_find(aa, 0)));
 
     double * va = emalloc(m * n * sizeof(double));
     int * pivots = emalloc(MIN(m, n) * sizeof(int));
 
     for (i = 0; i < m; ++i) {
-        row = Z_ARR(ba[i].val)->arData;
+        row = zend_hash_index_find(aa, i);
 
         for (j = 0; j < n; ++j) {
-            va[i * n + j] = zephir_get_doubleval(&row[j].val);
+            va[i * n + j] = zephir_get_doubleval(zend_hash_index_find(Z_ARR_P(row), j));
         }
     }
 
@@ -306,22 +294,20 @@ void tensor_ref(zval * return_value, zval * a)
 void tensor_cholesky(zval * return_value, zval * a)
 {
     unsigned int i, j;
-    Bucket * row;
+    zval * row;
     zval rowB, b;
 
     zend_array * aa = Z_ARR_P(a);
-
-    Bucket * ba = aa->arData;
 
     unsigned int n = zend_array_count(aa);
 
     double * va = emalloc(n * n * sizeof(double));
 
     for (i = 0; i < n; ++i) {
-        row = Z_ARR(ba[i].val)->arData;
+        row = zend_hash_index_find(aa, i);
 
         for (j = 0; j < n; ++j) {
-            va[i * n + j] = zephir_get_doubleval(&row[j].val);
+            va[i * n + j] = zephir_get_doubleval(zend_hash_index_find(Z_ARR_P(row), j));
         }
     }
 
@@ -361,13 +347,11 @@ void tensor_cholesky(zval * return_value, zval * a)
 void tensor_lu(zval * return_value, zval * a)
 {
     unsigned int i, j;
-    Bucket * row;
+    zval * row;
     zval rowL, l, rowU, u, rowP, p;
     zval tuple;
 
     zend_array * aa = Z_ARR_P(a);
-
-    Bucket * ba = aa->arData;
 
     unsigned int n = zend_array_count(aa);
 
@@ -375,10 +359,10 @@ void tensor_lu(zval * return_value, zval * a)
     int * pivots = emalloc(n * sizeof(int));
 
     for (i = 0; i < n; ++i) {
-        row = Z_ARR(ba[i].val)->arData;
+        row = zend_hash_index_find(aa, i);
 
         for (j = 0; j < n; ++j) {
-            va[i * n + j] = zephir_get_doubleval(&row[j].val);
+            va[i * n + j] = zephir_get_doubleval(zend_hash_index_find(Z_ARR_P(row), j));
         }
     }
 
@@ -457,15 +441,13 @@ void tensor_lu(zval * return_value, zval * a)
 void tensor_eig(zval * return_value, zval * a)
 {
     unsigned int i, j;
-    Bucket * row;
+    zval * row;
     zval eigenvalues;
     zval eigenvectors;
     zval eigenvector;
     zval tuple;
 
     zend_array * aa = Z_ARR_P(a);
-
-    Bucket * ba = aa->arData;
 
     unsigned int n = zend_array_count(aa);
 
@@ -475,10 +457,10 @@ void tensor_eig(zval * return_value, zval * a)
     double * vr = emalloc(n * n * sizeof(double));
 
     for (i = 0; i < n; ++i) {
-        row = Z_ARR(ba[i].val)->arData;
+        row = zend_hash_index_find(aa, i);
 
         for (j = 0; j < n; ++j) {
-            va[i * n + j] = zephir_get_doubleval(&row[j].val);
+            va[i * n + j] = zephir_get_doubleval(zend_hash_index_find(Z_ARR_P(row), j));
         }
     }
 
@@ -525,7 +507,7 @@ void tensor_eig(zval * return_value, zval * a)
 void tensor_eig_symmetric(zval * return_value, zval * a)
 {
     unsigned int i, j;
-    Bucket * row;
+    zval * row;
     zval eigenvalues;
     zval eigenvectors;
     zval eigenvector;
@@ -533,18 +515,16 @@ void tensor_eig_symmetric(zval * return_value, zval * a)
 
     zend_array * aa = Z_ARR_P(a);
 
-    Bucket * ba = aa->arData;
-
     unsigned int n = zend_array_count(aa);
 
     double * va = emalloc(n * n * sizeof(double));
     double * wr = emalloc(n * sizeof(double));
 
     for (i = 0; i < n; ++i) {
-        row = Z_ARR(ba[i].val)->arData;
+        row = zend_hash_index_find(aa, i);
 
         for (j = 0; j < n; ++j) {
-            va[i * n + j] = zephir_get_doubleval(&row[j].val);
+            va[i * n + j] = zephir_get_doubleval(zend_hash_index_find(Z_ARR_P(row), j));
         }
     }
 
@@ -589,7 +569,7 @@ void tensor_eig_symmetric(zval * return_value, zval * a)
 void tensor_svd(zval * return_value, zval * a)
 {
     unsigned int i, j;
-    Bucket * row;
+    zval * row;
     zval u, rowU;
     zval s;
     zval vt, rowVt;
@@ -597,10 +577,8 @@ void tensor_svd(zval * return_value, zval * a)
 
     zend_array * aa = Z_ARR_P(a);
 
-    Bucket * ba = aa->arData;
-
     unsigned int m = zend_array_count(aa);
-    unsigned int n = zend_array_count(Z_ARR(ba[0].val));
+    unsigned int n = zend_array_count(Z_ARR_P(zend_hash_index_find(aa, 0)));
     unsigned int k = MIN(m, n);
 
     double * va = emalloc(m * n * sizeof(double));
@@ -609,10 +587,10 @@ void tensor_svd(zval * return_value, zval * a)
     double * vvt = emalloc(n * n * sizeof(double));
 
     for (i = 0; i < m; ++i) {
-        row = Z_ARR(ba[i].val)->arData;
+        row = zend_hash_index_find(aa, i);
 
         for (j = 0; j < n; ++j) {
-            va[i * n + j] = zephir_get_doubleval(&row[j].val);
+            va[i * n + j] = zephir_get_doubleval(zend_hash_index_find(Z_ARR_P(row), j));
         }
     }