--- a/ext/tensor/matrix.zep.h
+++ b/ext/tensor/matrix.zep.h
@@ -727,7 +727,7 @@ ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_tensor_matrix_offsetunset, 0, 1,
 	ZEND_ARG_INFO(0, index)
 ZEND_END_ARG_INFO()
 
-ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_tensor_matrix_offsetget, 0, 1, IS_ARRAY, 0)
+ZEND_BEGIN_ARG_WITH_RETURN_OBJ_INFO_EX(arginfo_tensor_matrix_offsetget, 0, 1, Tensor\\Vector, 0)
 	ZEND_ARG_INFO(0, index)
 ZEND_END_ARG_INFO()