… tensor/matrix.zep
… tensor/vector.zep
… tests/MatrixTest.php
… tests/VectorTest.php