This PR changes how arrays are accessed in all the operations. It appears that for some reason the hashtables are becoming non-packed, but I am not sure why. Changing directly accessing `arData` to using `zend_hash_index_find` fixes this issue. Under the hood `zend_hash_index_find` evaluates whether the array is properly packed, and delegates to directly accessing arData. So I think this change should be as performant as it was before, if your arrays are properly packed.

All the tests run successfully with these changes.

There are still deprecation warnings about `self` and `static` calls, but it looks like Zephir has not yet developed a fix for this. https://github.com/zephir-lang/zephir/issues/2405

There are also warnings like "Implicit conversion from float 3.3 to int loses precision", which appear to be coming from calls into Zend mod_function. I was not sure of the mathematically correct way to resolve this one.
Minor changes like `int` to `float` warnings are also fixed, and a return type of `Vector` is fixed to match the non-extension code.

(Potentially) fixes #33 and #38