@andrewdalpino 2023-09-20
Awesome @27pchrisl thanks for looking into this for us! Your help is greatly appreciated.

Did you run the benchmarks with these changes yet by any chance? I'm curious to know if we take a performance hit. If there is a performance hit, do you think we could try to find the point at which these arrays become "unpacked" and "re-pack" them as a way to mitigate the performance loss?