@andrewdalpino 2023-09-20
There doesn't appear to be much (if any) of a performance impact.

Baseline ...

| MatrixMatrixMultiplyBench       | multiply     |     | 1    | 5   | 112.114mb | 0.015s   | ±9.78%  |
| MatrixVectorMultiplyBench       | multiply     |     | 1    | 5   | 75.193mb  | 0.011s   | ±9.72%  |
| MatrixScalarMultiplyBench       | multiply     |     | 1    | 5   | 75.119mb  | 0.011s   | ±1.85%  |
| VectorVectorMultiplyBench       | multiply     |     | 1    | 5   | 3.737mb   | 0.385ms  | ±2.14%  |

After your changes ...

| MatrixMatrixMultiplyBench       | multiply     |     | 1    | 5   | 112.114mb | 0.015s   | ±12.67% |
| MatrixVectorMultiplyBench       | multiply     |     | 1    | 5   | 75.193mb  | 0.012s   | ±9.72%  |
| MatrixScalarMultiplyBench       | multiply     |     | 1    | 5   | 75.119mb  | 0.011s   | ±0.19%  |
| VectorVectorMultiplyBench       | multiply     |     | 1    | 5   | 3.737mb   | 0.393ms  | ±58.90% |